fix(codec): the header is a fixed width and usize is off the wire
A collection's header was two native `usize`, so it was sixteen bytes where a
pointer is eight and eight where it is four, and `usize`/`isize` fields carried
the same split into the payload.

`encrypt_into_buffers` never asks for `encode_bytes_required` — `get_sizes`
does, one level up — so the injection its test named was inert, and what the
test pinned was the buffer it overflowed. It now says so.

+++ crates/redoubt-codec/core/src/tests/collections/redoubt_vec.rs
@@ -43,7 +43,7 @@ fn stress_test_redoubt_vec_clear_push_encode_decode_cycles()
 
     // Create original array (clonable)
     let original: Vec<RedoubtCodecTestBreaker> = (0..SIZE)
-        .map(|i| RedoubtCodecTestBreaker::new(RedoubtCodecTestBreakerBehaviour::None, i))
+        .map(|i| RedoubtCodecTestBreaker::new(RedoubtCodecTestBreakerBehaviour::None, i as u32))
         .collect();
 
     let mut redoubt_vec = RedoubtVec::<RedoubtCodecTestBreaker>::new();
